+++ c/vector.h
@@ -1,10 +1,11 @@
+#include <stdint.h>
 #ifndef _VECTORH_
 #define _VECTORH_
 
 typedef struct Vector {
   int capacity;
   int total;
-  int * items;
+  int8_t * items;
 } Vector;
 
 #define UNDEFINE -1
@@ -12,9 +13,9 @@ typedef struct Vector {
 
 int vectorTotal(Vector *v);
 int vectorResize(Vector *v, int capacity);
-int vectorAppend(Vector *v, int item);
-int vectorSet(Vector *v, int index, int item);
-int vectorGet(Vector *v, int index);
+int vectorAppend(Vector *v, int8_t item);
+int vectorSet(Vector *v, int index, int8_t item);
+int8_t vectorGet(Vector *v, int index);
 int vectorDelete(Vector *v, int index);
 int vectorFree(Vector *v);
 void vector_init(Vector *v, int capacity);
